Add TS definition 3.0 features to definition file
It's now possible to resolve Array types from argument types. Thought that dependency injection could use some of this resolution. This is what I have in my local. Let me know if my assumptions about keys are incorrect. In my usage I used base constructors as the keys, and then instances which implement the contract of the constructor.